--- src/main/java/com/dev/cinema/controllers/HelloController.java
@@ -1,13 +1,0 @@
-package com.dev.cinema.controllers;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class HelloController {
-
-    @GetMapping("/hello")
-    public String sayHello() {
-        return "Hello there!";
-    }
-}
