+++ src/main/java/com/dev/cinema/controllers/mappers/MovieSessionMapper.java
@@ -1,4 +1,4 @@
-package com.dev.cinema.controllers;
+package com.dev.cinema.controllers.mappers;
 
 import com.dev.cinema.model.MovieSession;
 import com.dev.cinema.model.dto.MovieSessionRequestDto;
