+++ src/main/java/com/dev/cinema/model/dto/UserRequestDto.java
@@ -1,7 +1,6 @@
 package com.dev.cinema.model.dto;
 
 public class UserRequestDto {
-    private Long id;
     private String email;
     private String password;
     private String repeatedPassword;
@@ -9,19 +8,10 @@ public class UserRequestDto {
     public UserRequestDto() {
     }
 
-    public UserRequestDto(Long id, String email, String password, String repeatedPassword) {
-        this.id = id;
+    public UserRequestDto(String email, String password, String repeatedPassword) {
         this.email = email;
         this.password = password;
         this.repeatedPassword = repeatedPassword;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
     }
 
     public String getEmail() {
